Merge Client.Tests and Server.Tests into single..
... test project (Presentation.Tests).

This solves the code coverage error which was being generated. The reason why that error came up was because the Server.dll assembly depends on the Client.dll assembly and so when tests are run in Server.Tests.dll it only detects tests on Server.dll but it sees that there's an untested Client.dll and so it counts it in the coverage. Since those two projects are thighly coupled it is reasonable to keep them in the same test project.